Call setupQuizHTML from initializeQuiz
When starting the quiz the page is not loaded because the setup is not actually called: the initialiser `initializeQuiz` is called, but it does not call the `setupQuizHTML` function. 
Because of this nothing is shown. I thought this might be something for the students to figure out (and was missing intentionally). However no one seems to be able when this far in the curriculum, to find this out for themselves: getting used to multiple files and importing and exporting is already a (big) challenge.